+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 * Support for specifying the MIME type when embedding files
+* Support for adding custom XMP metadata
 
 
 ## 0.43.0 - 2024-05-26